requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.7
pillow==12.1.0
pycparser==3.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
